requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
agno==2.2.6
    # via message-sender-microsservice (pyproject.toml)
amqp==5.3.1
    # via kombu
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==5.0.0
    # via message-sender-microsservice (pyproject.toml)
billiard==4.2.2
    # via celery
cachetools==6.2.1
    # via google-auth
celery==5.5.3
    # via message-sender-microsservice (pyproject.toml)
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   rich-toolkit
    #   typer
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cryptography==46.0.3
    # via python-jose
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via agno
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.120.4
    # via message-sender-microsservice (pyproject.toml)
fastapi-cli==0.0.14
    # via fastapi
fastapi-cloud-cli==0.3.1
    # via fastapi-cli
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via agno
google-auth==2.42.1
    # via google-genai
google-genai==1.47.0
    # via message-sender-microsservice (pyproject.toml)
h11==0.16.0
    # via
    #   agno
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   agno
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
kombu==5.5.4
    # via celery
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.3.4
    # via
    #   pandas
    #   pgvector
orjson==3.11.4
    # via fastapi
packaging==25.0
    # via
    #   agno
    #   kombu
    #   pytest
pandas==2.3.3
    # via message-sender-microsservice (pyproject.toml)
pgvector==0.4.1
    # via message-sender-microsservice (pyproject.toml)
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via message-sender-microsservice (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   message-sender-microsservice (pyproject.toml)
    #   agno
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-extra-types==2.10.6
    # via fastapi
pydantic-settings==2.11.0
    # via
    #   message-sender-microsservice (pyproject.toml)
    #   agno
    #   fastapi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==9.0.1
    # via message-sender-microsservice (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   pandas
python-dotenv==1.2.1
    # via
    #   agno
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via message-sender-microsservice (pyproject.toml)
python-multipart==0.0.20
    # via
    #   agno
    #   fastapi
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   agno
    #   fastapi
    #   uvicorn
redis==7.0.1
    # via message-sender-microsservice (pyproject.toml)
requests==2.32.5
    # via google-genai
rich==14.2.0
    # via
    #   agno
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.3
    # via fastapi-cloud-cli
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
sentry-sdk==2.43.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
starlette==0.49.3
    # via fastapi
tenacity==9.1.2
    # via google-genai
typer==0.20.0
    # via
    #   agno
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   agno
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via
    #   kombu
    #   pandas
ujson==5.11.0
    # via fastapi
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.38.0
    # via
    #   message-sender-microsservice (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
